abstract iammeterBaseHandler
